requirements.txt
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.23.5
soundfile==0.12.1
fastapi==0.111.0
uvicorn==0.30.1
demucs==4.0.1
basic-pitch==0.3.1
pyjwt==2.8.0
loguru==0.7.2
matplotlib==3.9.0
mido==1.3.3
music21==8.3.0
python-dotenv==1.0.1
sentencepiece==0.2.0
transformers==4.38.2
speechbrain==1.0.1
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
einops==0.8.0
audiocraft==1.3.0
protobuf==3.20.3
scikit-learn==1.3.2
